requirements.txt
requests>=2.28.0
curl_cffi>=0.5.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dotenv>=1.0.0
pyyaml>=6.0
schedule>=1.2.0
flask>=3.0.0
click>=8.1.0
anthropic>=0.25.0
pytesseract>=0.3.10
Pillow>=10.0.0
